Refactor Employee and Project Management Logic
- Moved user projects logic from ProjectsController to EmployeeProjectsController for better separation of concerns.
- Centralized project access validation in EmployeeProjectsController.
- Created EmployeeAuthentication, EmployeeFactory, EmployeeFilter, and EmployeePaginator classes to encapsulate related functionalities.
- Updated routes to point user projects to the new method in EmployeeProjectsController.
- Enhanced project access checks and employee role retrieval methods.
- Added a new view for displaying user-specific project details.
