+++ app/Controllers/EmployeeProjectsController.php
@@ -18,6 +18,7 @@ class EmployeeProjectsController extends Controller
     private const ACCESS_DENIED = 'Access denied';
     private const ASSIGNMENT_CREATED = 'Employee assigned to project successfully!';
     private const ASSIGNMENT_REMOVED = 'Employee removed from project successfully!';
+    private const MY_PROJECTS = 'My Projects';
 
     public function __construct()
     {
@@ -25,9 +26,6 @@ class EmployeeProjectsController extends Controller
 
         if (!Auth::check()) {
             $this->redirectTo(route('auth.login'));
-        } elseif (!Auth::isHR() && !Auth::isAdmin()) {
-            FlashMessage::danger(self::ACCESS_DENIED);
-            $this->redirectTo(route('user.home'));
         }
     }
 
@@ -55,6 +53,12 @@ class EmployeeProjectsController extends Controller
             }
 
           // Check if the employee is already assigned to the project
+            if (!($project instanceof Project) || !method_exists($project, 'employees')) {
+                FlashMessage::danger('Erro: Projeto inválido');
+                $this->redirectTo(route('projects.index'));
+                return;
+            }
+
             $existingAssignments = $project->employees()->get();
             foreach ($existingAssignments as $existingEmployee) {
                 if ($existingEmployee->id === $employeeId) {
@@ -65,6 +69,12 @@ class EmployeeProjectsController extends Controller
             }
 
           // Assign employee to project with role
+            if (!($project instanceof Project) || !method_exists($project, 'employees')) {
+                FlashMessage::danger('Erro: Projeto inválido');
+                $this->redirectTo(route('projects.index'));
+                return;
+            }
+
             $project->employees()->attach($employeeId, ['role' => $role]);
 
             FlashMessage::success(self::ASSIGNMENT_CREATED);
@@ -98,6 +108,12 @@ class EmployeeProjectsController extends Controller
             }
 
           // Remove employee from project
+            if (!($project instanceof Project) || !method_exists($project, 'employees')) {
+                FlashMessage::danger('Erro: Projeto inválido');
+                $this->redirectTo(route('projects.index'));
+                return;
+            }
+
             $project->employees()->detach($employeeId);
 
             FlashMessage::success(self::ASSIGNMENT_REMOVED);
@@ -118,10 +134,132 @@ class EmployeeProjectsController extends Controller
             return;
         }
 
-      // Get projects associated with this employee
+        // Get projects associated with this employee
         $projects = $employee->projects()->get();
 
         $title = 'Projects for ' . $employee->name;
         $this->render('employee_projects/index', compact('employee', 'projects', 'title'));
+    }
+
+    /**
+     * Lista os projetos associados ao usuário atual
+     * Método movido do ProjectsController para centralizar a lógica
+     */
+    public function userProjects(): void
+    {
+        if (!Auth::check()) {
+            $this->redirectTo(route('auth.login'));
+            return;
+        }
+
+        $currentUser = Auth::user();
+
+        if (!$currentUser) {
+            FlashMessage::danger('Usuário não encontrado');
+            $this->redirectTo(route('auth.login'));
+            return;
+        }
+
+        // Obter o funcionário associado ao usuário atual
+        $employee = Employee::getCurrentUserEmployee();
+
+        if (!$employee) {
+            FlashMessage::danger('Funcionário não encontrado');
+            $this->redirectTo(route('user.home'));
+            return;
+        }
+
+        // Obter os projetos associados ao funcionário atual
+        $userProjects = $employee->projects()->get();
+
+        // Obter informações adicionais sobre cada projeto
+        $projectsWithDetails = [];
+        foreach ($userProjects as $project) {
+            // Garantir que o projeto é uma instância válida
+            if (!($project instanceof Project) || !method_exists($project, 'employees')) {
+                // Se não for um projeto válido, pule para o próximo
+                continue;
+            }
+
+            $projectEmployees = $project->employees()->get();
+            $employeeRoles = $this->getEmployeeProjectRoles($project->id);
+
+            $projectsWithDetails[] = [
+                'project' => $project,
+                'role' => $employeeRoles[$employee->id] ?? 'Não especificado',
+                'team_size' => count($projectEmployees)
+            ];
+        }
+
+        $title = self::MY_PROJECTS;
+        $this->render('projects/user_projects', compact('projectsWithDetails', 'title'));
+    }
+
+    /**
+     * Verifica se um usuário tem acesso a um projeto específico
+     * Método centralizado para validação de acesso
+     *
+     * @param int $projectId ID do projeto
+     * @return bool True se o usuário tem acesso, false caso contrário
+     */
+    public function hasProjectAccess(int $projectId): bool
+    {
+        // Admin e HR sempre têm acesso
+        if (Auth::isAdmin() || Auth::isHR()) {
+            return true;
+        }
+
+        $hasAccess = false;
+        $employee = Employee::getCurrentUserEmployee();
+
+        // Verifica se o funcionário existe e se o projeto está na lista de projetos dele
+        if ($employee) {
+            // Obtém todos os projetos do funcionário atual
+            $employeeProjects = $employee->projects()->get();
+
+            // Verifica se o projeto está na lista de projetos do funcionário
+            foreach ($employeeProjects as $project) {
+                if ((int)$project->id === (int)$projectId) {
+                    $hasAccess = true;
+                    break;
+                }
+            }
+        }
+
+        return $hasAccess;
+    }
+
+    /**
+     * Retorna o papel de cada funcionário em um projeto específico
+     * Método movido do ProjectsController para centralizar a lógica
+     *
+     * @param int $projectId ID do projeto
+     * @return array Array associativo com [employee_id => role]
+     */
+    public function getEmployeeProjectRoles(int $projectId): array
+    {
+        $project = Project::findById($projectId);
+
+        if (!$project) {
+            return [];
+        }
+
+        $roles = [];
+        try {
+            $pdo = \Core\Database\Database::getDatabaseConn();
+            $query = "SELECT employee_id, role FROM employee_projects WHERE project_id = :project_id";
+            $stmt = $pdo->prepare($query);
+            $stmt->bindValue(':project_id', $projectId);
+            $stmt->execute();
+            $results = $stmt->fetchAll(\PDO::FETCH_ASSOC);
+
+            foreach ($results as $result) {
+                $roles[$result['employee_id']] = $result['role'] ?? 'Membro da equipe';
+            }
+        } catch (\Exception $e) {
+            // Log error and continue with empty roles array
+        }
+
+        return $roles;
     }
 }
